Adding a version for the language, and include a new rev (#118)
Following up on #116 - now that there is a release/tag with the pre-commit-hooks.yaml, I believe it should be included in the README!

Also included is a line for recommending the node language_version of 18.19.1 (latest, stable for node 18). Could be updated to 20 if you'd like. I added this because my own dev environment got a little wonky and messed up the initial pre-commit setup until this was explicitly called out.